move helper classes nested in instrumentation: letter A
