+++ dd-java-agent/instrumentation/apache-httpclient-4/src/main/java/datadog/trace/instrumentation/apachehttpclient/HelperMethods.java
@@ -32,7 +32,7 @@ public class HelperMethods {
   }
 
   public static void doMethodExit(
-    final AgentScope scope, final Object result, final Throwable throwable) {
+      final AgentScope scope, final Object result, final Throwable throwable) {
     if (scope == null) {
       return;
     }
