+++ dd-smoke-tests/agent-isolation/src/main/java/datadog/smoketest/agentisolation/App.java
@@ -1,6 +1,5 @@
 package datadog.smoketest.agentisolation;
 
 public class App {
-  public static void main(String... args) {
-  }
+  public static void main(String... args) {}
 }
